Fix: Angular implementation, new config components, lint, and build
