--- Angular/src/app/orig_app.component.scss
@@ -1,5 +1,0 @@
-.demo-container {
-  margin: 50px;
-  width: 90vh;
-}
-
