--- Angular/src/app/orig_app.module.ts
@@ -1,20 +1,0 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import { DxDataGridModule, DxButtonModule } from 'devextreme-angular';
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-
-@NgModule({
-  declarations: [
-    AppComponent,
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    DxDataGridModule,
-    DxButtonModule,
-  ],
-  providers: [],
-  bootstrap: [AppComponent],
-})
-export class AppModule { }
