+++ jQuery/package.json
@@ -5,9 +5,9 @@
     "start": "npm run lite",
     "build": "npm run lint",
     "lite": "lite-server -c ./bs-config.json",
-    "lint-html": "prettier --write src/*",
-    "lint-js": "eslint --ext .js . --fix",
-    "lint-css": "stylelint src/**/*.css --allow-empty-input --fix",
+    "lint-html": "prettier --check src/*",
+    "lint-js": "eslint --ext .js .",
+    "lint-css": "stylelint src/**/*.css --allow-empty-input",
     "lint": "npm-run-all -p -c lint-js lint-css lint-html"
   },
   "devDependencies": {
